src/main/java/com/tteam/movieland/util/Utils.java
package com.tteam.movieland.util;

import com.tteam.movieland.exception.RandomMoviesException;

import java.util.Collections;
import java.util.List;
import java.util.Random;
import java.util.concurrent.ThreadLocalRandom;

public class Utils {
    public static <E> List<E> pickNRandomElements(List<E> list, int n, Random r) {
        int length = list.size();

        if (length < n) throw new RandomMoviesException(String.format("""
        Cannot create list of movies by provided number. Number is %1$s, total amount of movies is %2$s
        """, n, length));

        //We don't need to shuffle the whole list
        for (int i = length - 1; i >= length - n; --i)
        {
            Collections.swap(list, i , r.nextInt(i + 1));
        }
        return list.subList(length - n, length);
    }

    public static <E> List<E> pickNRandomElements(List<E> list, int n) {
        return pickNRandomElements(list, n, ThreadLocalRandom.current());
    }
}
